+++ packages/components/src/components/segmented-control/segmented-control/segmented-control.spec.ts
@@ -68,18 +68,6 @@ describe('render', () => {
 
     expect(spy).toHaveBeenCalledWith(component.host, component.value, component.disabled, component.theme);
   });
-
-  it('should call warnIfDeprecatedPropIsUsed() with correct parameters', () => {
-    const spy = jest.spyOn(warnIfDeprecatedPropIsUsed, 'warnIfDeprecatedPropIsUsed');
-    const component = initComponent();
-    component.host = document.createElement('p-segmented-control');
-    component.backgroundColor = 'background-surface';
-    component.host.attachShadow({ mode: 'open' });
-
-    component.render();
-
-    expect(spy).toHaveBeenCalledWith(component, 'backgroundColor');
-  });
 });
 
 describe('updateSegmentedControlItemHandler', () => {